requirements.txt
ape-foundry==0.6.15
ape-vyper==0.6.10
eth-ape==0.6.19
hypothesis>=6.82.4
